+++ src/olmo_core/data/__init__.py
@@ -6,8 +6,10 @@ Overview
 
 For text-based data you should prepare your data by writing token IDs to numpy arrays on disk, using the
 `Dolma toolkit <https://allenai.github.io/dolma/>`_ for example.
-Then configure and build your dataset using the :class:`~olmo_core.data.numpy_dataset.NumpyDatasetConfig`
-builder, build your data loader with the :class:`~olmo_core.data.data_loader.NumpyDataLoaderConfig`
+Then configure and build your dataset using one of the
+:class:`~olmo_core.data.numpy_dataset.NumpyDatasetConfigBase` builders (for example
+``NumpyFSLDatasetConfig``), build your data loader with the
+:class:`~olmo_core.data.data_loader.NumpyDataLoaderConfig`
 builder, and pass it to :meth:`TrainerConfig.build() <olmo_core.train.TrainerConfig.build>`.
 """
 
@@ -24,12 +26,17 @@ from .mixes import DataMix, DataMixBase
 from .numpy_dataset import (
     InstanceFilterConfig,
     NumpyDatasetBase,
-    NumpyDatasetConfig,
+    NumpyDatasetConfigBase,
     NumpyFSLDataset,
     NumpyFSLDatasetBase,
+    NumpyFSLDatasetConfig,
+    NumpyInterleavedFSLDatasetConfig,
     NumpyPackedFSLDataset,
+    NumpyPackedFSLDatasetConfig,
     NumpyPaddedFSLDataset,
+    NumpyPaddedFSLDatasetConfig,
     NumpyVSLDataset,
+    NumpyVSLDatasetConfig,
     VSLCurriculum,
     VSLCurriculumConfig,
     VSLCurriculumType,
@@ -39,7 +46,7 @@ from .numpy_dataset import (
     VSLNaturalCurriculum,
 )
 from .tokenizer import TokenizerConfig, TokenizerName
-from .types import LongDocStrategy, NumpyDatasetDType, NumpyDatasetType
+from .types import LongDocStrategy, NumpyDatasetDType
 
 __all__ = [
     "NumpyDatasetBase",
@@ -53,8 +60,12 @@ __all__ = [
     "VSLGrowthCurriculum",
     "VSLGrowP2Curriculum",
     "VSLGrowLinearCurriculum",
-    "NumpyDatasetConfig",
-    "NumpyDatasetType",
+    "NumpyDatasetConfigBase",
+    "NumpyFSLDatasetConfig",
+    "NumpyPaddedFSLDatasetConfig",
+    "NumpyPackedFSLDatasetConfig",
+    "NumpyInterleavedFSLDatasetConfig",
+    "NumpyVSLDatasetConfig",
     "InstanceFilterConfig",
     "VSLCurriculumType",
     "VSLCurriculumConfig",