+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `BeakerLaunchConfig.launch()` will now send Slack notifications by default when `follow=True` if the env var `SLACK_WEBHOOK_URL` is set.
 - `src/examples/llama/` has been renamed to `src/examples/llm/`.
 - Refactored eval task groups into `task_groups.py`
+- Refactored `NumpyDatasetConfig` by splitting it into a separate config per underlying dataset class.
 
 ### Added